+++ pyodide_pack/runtime_detection.py
@@ -9,7 +9,8 @@ from pathlib import Path
 from pyodide_pack._utils import (
     match_suffix,
 )
-from pyodide_pack.archive import ArchiveFile
+from pyodide_pack.ast_rewrite import _rewrite_py_code
+from pyodide_pack.config import PackConfig
 from pyodide_pack.dynamic_lib import DynamicLib
 
 
@@ -88,8 +89,9 @@ class RuntimeResults(dict):
 class PackageBundler:
     """Only include necessary files for a given package."""
 
-    def __init__(self, db: RuntimeResults, include_paths: str | None = None):
+    def __init__(self, db: RuntimeResults, config: PackConfig):
         self.db = db
+        self.config = config
         self.stats = {
             "py_in": 0,
             "so_in": 0,
@@ -102,7 +104,6 @@ class PackageBundler:
             "size_gzip_out": 0,
         }
         self.dynamic_libs: list[DynamicLib] = []
-        self.include_paths = include_paths
 
     def process_path(self, in_file_name: str) -> str | None:
         """Process a path, returning the output path if it should be included."""
@@ -135,9 +136,9 @@ class PackageBundler:
                 case _:
                     stats["other_out"] += 1
 
-        elif self.include_paths is not None and any(
+        elif self.config.include_paths is not None and any(
             fnmatch.fnmatch(in_file_name, pattern)
-            for pattern in self.include_paths.split(",")
+            for pattern in self.config.include_paths
         ):
             # TODO: this is hack and should be done better
             out_file_name = os.path.join("/lib/python3.11/site-utils", in_file_name)
@@ -151,21 +152,18 @@ class PackageBundler:
                     self.dynamic_libs.append(dll)
         return out_file_name
 
-    def copy_between_zip_files(
-        self,
-        in_file_name: str,
-        out_file_name: str,
-        archive: ArchiveFile,
-        fh_out,
-    ) -> None:
-        """Copy a file between two archives."""
+    def process_content(self, in_file_name: str, content: bytes) -> bytes | None:
+        """Process both the input filename and the file contents"""
         stats = self.stats
         stats["fh_out"] += 1
-        stream = archive.read(in_file_name)
-        if stream is not None:
-            stats["size_out"] += len(stream)
-            stats["size_gzip_out"] += len(gzip.compress(stream))
-            # File paths starting with / fails to get correctly extracted
-            # in extract_archive in Pyodide
-            with fh_out.open(out_file_name.lstrip("/"), "w") as fh:
-                fh.write(stream)
+
+        extension = Path(in_file_name).suffix
+        out_content = content
+        if extension == ".py":
+            out_content = _rewrite_py_code(
+                content.decode(), file_name=in_file_name, py_config=self.config.py
+            ).encode()
+        stats["size_out"] += len(out_content)
+        stats["size_gzip_out"] += len(gzip.compress(out_content))
+
+        return out_content