+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+ - Add configuration for `pyodide pack` loaded from the `[tool.pyodide_pack]` section of a
+   `pyproject.toml` files
+   [#35](https://github.com/pyodide/pyodide-pack/pull/35)
+
+
  - Add `pyidide minify` command to minify the Python packages with AST rewrites by,
    removing comments and docstrings
    [#23](https://github.com/pyodide/pyodide-pack/pull/23)
@@ -28,6 +33,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - Added support for Pyodide 0.24.0. This is now the minimal supported version of Pyodide.
    [#26](https://github.com/pyodide/pyodide-pack/pull/26)
+
+## Fixed
+
+ - Fix a syntax error when stripping docstrings from a function with an empty body
+   [#35](https://github.com/pyodide/pyodide-pack/pull/35)
+
 
 ## [0.2.0] - 2022-09-04